refactor: delete the runner
